Make core minisat produce very verbose trace.
Output all added and removed clauses with identifiers.
Output all branches, all backtracks, and all
implications and conflicts with antecedants.

Also turn off ccmin by default as it may depend
on learned clauses in a nontrivial way which
I have not decided how to deal with yet.
I will turn it back on in the future.