Allow minisat to use decision oracle.
Add an optional command line argument for core minisat to
take the filename of the analysis output and use it as
a source of decisions instead of picking itself.

When using this decision oracle, the normal reset
mechanism is turned off so that resets only occur as given
by the oracle.