Make BCP more deterministic.
If multiple clauses become unit simultaneously
and imply the same literal, make sure the one with the
highest ID is the one recorded as the reason.

If multiple clauses become unit simultaneously and imply
contradictory literals, make sure the one with the unnegated
literal is analyzed first as an implication and the one with
the negated literal is the final clause analyzed as the source
of the conflict.

If multiple clauses become UNSAT simultaneously, make sure the one
with the highest ID is the conflict that is analyzed.